# auto-center-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AC API",
        "description": "API for Startup",
        "contact": {
            "name": "Farhad Haider",
            "email": "farhadandproject@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/get-token": {
            "post": {
                "tags": [
                    "Get Token"
                ],
                "summary": "User login",
                "description": "Logs a user into the system and returns an access token",
                "operationId": "login",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "your_access_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/vehicle/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get a specific vehicle by ID",
                "description": "Retrieve detailed information about a specific vehicle.",
                "operationId": "getVehicleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vehicle to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/vehicles": {
            "get": {
                "tags": [
                    "All Vehicles"
                ],
                "summary": "Get all vehicles",
                "description": "Retrieve a list of all vehicles with detailed information.",
                "operationId": "getAllVehicles",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vehicle"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/search": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Search vehicles",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "q": {
                                        "description": "Search query",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Result not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Vehicle": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "make": {
                        "type": "string"
                    },
                    "model": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "drive_type": {
                        "type": "string"
                    },
                    "transmission": {
                        "type": "string"
                    },
                    "condition": {
                        "type": "string"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "mileage": {
                        "type": "integer"
                    },
                    "fuel_type": {
                        "type": "string"
                    },
                    "engine_size": {
                        "type": "string"
                    },
                    "doors": {
                        "type": "integer"
                    },
                    "cylinders": {
                        "type": "integer"
                    },
                    "VIN": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "stock_id": {
                        "type": "string"
                    },
                    "discount": {
                        "type": "number"
                    },
                    "is_featured": {
                        "type": "boolean"
                    },
                    "featured_from_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "featured_to_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "dealership_id": {
                        "type": "integer"
                    },
                    "is_sold": {
                        "type": "boolean"
                    },
                    "is_enabled": {
                        "type": "boolean"
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "rating": {
                        "type": "number"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}